Reduce response payload of Calendar API.
Also minor changes for inlining code and and
reducing await calls.

Previously, the response payload was quite large
as all information was being fetched. We only need
a couple of fields, so we are now only fetching
those. This should reduce the payload size, so
reduces the time it takes to fetch the data, so
reduces the time it takes to SSR the page.

Also, when we hit a cache miss, we still return
the stale results now, but load for the next page
refresh. Stale results are fine 99.9% of the time,
as the data doesn't change that often.